+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 ## Unreleased
 
+### Added
+- When removing links, remove cursor and restore scroll position.
+
 ### Fixed
 - When changing the displayed node, systematically remove focus to avoid keyboard/cursor inconsistencies.
 - The “Remove links” option is still enabled right after hiding the keyboard.
+
 
 ## v0.1.3
 ### Added